+++ src/renderer/light.h
@@ -6,22 +6,24 @@
 struct light_t
 {
     vec3    colour = { 1.f, 1.f, 1.f };
-    float   ambient_intensity = 0.2f;
     float   diffuse_intensity = 1.0f;
 };
 
 struct directional_light_t : light_t
 {
+    float   ambient_intensity = 0.2f;
     quaternion orientation = { 0.7071068f, 0.f, 0.f, 0.7071068f };
 };
 
 struct point_light_t : light_t
 {
     vec3        position = { 0.f, 0.f, 0.f };
+    float       radius = 0.f;
     // Attenuation coefficients
-    float       att_constant = 0.3f;
-    float       att_linear = 0.2f;
-    float       att_quadratic = 0.1f;
+    float       att_constant = 0.1f; //0.3f;
+    float       att_linear = 0.1f; //0.2f;
+    float       att_quadratic = 0.6f; //0.1f;
+    u32:32;
 };
 
 struct spot_light_t : point_light_t